refactor(desktop): extract runEnrichmentBatch service from metadata-enrich handler
